# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connectors/sync:
    post:
      operationId: syncConnector
      tags: [connectors]
      summary: Normalize a read-only provider sync batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectorSyncInput"
      responses:
        "201":
          description: Connector sync completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorSyncResult"
        "400":
          description: Invalid connector payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Write operation requires CerbaSeal authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /brain-versions:
    post:
      operationId: createBrainVersion
      tags: [brain-versions]
      summary: Create and verify a checksummed brain snapshot
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrainVersionInput"
      responses:
        "201":
          description: Brain snapshot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrainVersion"
    get:
      operationId: listBrainVersions
      tags: [brain-versions]
      summary: List brain snapshots
      responses:
        "200":
          description: Brain snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrainVersionSummary"
  /brain-versions/{id}/restore:
    post:
      operationId: restoreBrainVersion
      tags: [brain-versions]
      summary: Verify a snapshot and initiate a restore preflight
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreBrainVersionInput"
      responses:
        "200":
          description: Restore verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreBrainVersionResult"
        "404":
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ConnectorProvider:
      type: string
      enum: [gmail, proton, github, google_drive, google_calendar, replit]
    ConnectorEventInput:
      type: object
      properties:
        externalId:
          type: string
          minLength: 1
        eventType:
          type: string
          minLength: 1
        sourceRef:
          type: string
          minLength: 1
        occurredAt:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
      required: [externalId, eventType, sourceRef, occurredAt, payload]
    ConnectorSyncInput:
      type: object
      properties:
        provider:
          $ref: "#/components/schemas/ConnectorProvider"
        mode:
          type: string
          enum: [read, write]
          default: read
        events:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/ConnectorEventInput"
      required: [provider, events]
    ConnectorSyncResult:
      type: object
      properties:
        syncId:
          type: string
          format: uuid
        provider:
          $ref: "#/components/schemas/ConnectorProvider"
        status:
          type: string
        receivedCount:
          type: integer
        normalizedCount:
          type: integer
        eventIds:
          type: array
          items:
            type: string
            format: uuid
        domainEventId:
          type: string
          format: uuid
      required: [syncId, provider, status, receivedCount, normalizedCount, eventIds, domainEventId]
...
    CreateBrainVersionInput:
      type: object
      properties:
        versionName:
          type: string
          minLength: 1
          maxLength: 160
    BrainVersionSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        versionName:
          type: string
        schemaVersion:
          type: string
        status:
          type: string
        checksum:
          type: string
        totalRecords:
          type: integer
        createdAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
      required: [id, versionName, schemaVersion, status, checksum, totalRecords, createdAt]
    BrainVersion:
      allOf:
        - $ref: "#/components/schemas/BrainVersionSummary"
        - type: object
          properties:
            recordCounts:
              type: object
              additionalProperties:
                type: integer
            payload:
              type: object
              additionalProperties: true
          required: [recordCounts, payload]
    RestoreBrainVersionInput:
      type: object
      properties:
        confirm:
          type: boolean
          default: false
    RestoreBrainVersionResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        checksumValid:
          type: boolean
        status:
          type: string
        dryRun:
          type: boolean
        message:
          type: string
        eventId:
          type: string
          format: uuid
      required: [id, checksumValid, status, dryRun, message, eventId]
...